+++ avr/cores/megacommand/MCL/NoteInterface.cpp
@@ -155,7 +155,7 @@ void NoteInterface::draw_notes(uint8_t line_number) {
   /*Initialise the string with blank steps*/
   char str[17] = "----------------";
 
-  for (uint8_t i = 0; i < 16; i++) {
+  for (int i = 0; i < 16; i++) {
     if (notes[i] > 0 && notes[i] != 3) {
 
 #ifdef OLED_DISPLAY
